Added file response and upload functionality and removed the Blazor UI (#39)

+++ HashSlinger/Endpoints/HashtopolisApiV2/DTO/GetFileResponse.cs
@@ -13,7 +13,7 @@ public record GetFileResponse(
     string Extension,
     [property: JsonPropertyName("url")] string Url,
     [property: JsonPropertyName("filesize")]
-    int? FileSize,
+    long? FileSize,
     [property: JsonPropertyName("message"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     string? Message = null
 ) : IHashtopolisMessage;
